add hw estimates if detailed model is provided
